requirements.txt
python-crontab==3.2.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
google-auth==2.36.0
google-api-python-client==2.152.0
pytest-cov==6.0.0